feat(dolos): refactor and use common crate
