+++ crates/common/Cargo.toml
@@ -11,3 +11,9 @@ serde = { version = "1.0.218", features = ["derive"] }
 tracing = "0.1.41"
 serde_json = "1.0.135"
 reqwest = "0.12.12"
+bech32 = "0.9.1"
+cardano-serialization-lib = "12.1.1"
+hex = "0.4.3"
+blockfrost-openapi = "0.1.75"
+clap = { version = "4.5.26", features = ["derive"] }
+pallas-addresses = { git = "https://github.com/blockfrost/pallas.git", tag = "blockfrost-platform-0.0.3-alpha2" }
