+++ crates/common/src/lib.rs
@@ -1,3 +1,9 @@
+pub mod accounts;
+pub mod addresses;
+pub mod assets;
 pub mod errors;
+pub mod genesis;
 pub mod pagination;
+pub mod payment_cred;
+pub mod types;
 pub mod validation;
