+++ src/api/accounts/stake_address/root.rs
@@ -1,12 +1,8 @@
-use crate::{
-    BlockfrostError,
-    accounts::{AccountData, AccountsPath},
-    api::ApiResult,
-    server::state::AppState,
-};
+use crate::{BlockfrostError, api::ApiResult, server::state::AppState};
 
 use axum::extract::{Path, State};
 use blockfrost_openapi::models::account_content::AccountContent;
+use common::accounts::{AccountData, AccountsPath};
 
 pub async fn route(
     Path(path): Path<AccountsPath>,
