+++ src/api/accounts/stake_address/withdrawals.rs
@@ -1,12 +1,10 @@
-use crate::{
-    BlockfrostError,
-    accounts::{AccountData, AccountsPath},
-    api::ApiResult,
-    server::state::AppState,
-};
+use crate::{BlockfrostError, api::ApiResult, server::state::AppState};
 use axum::extract::{Path, Query, State};
 use blockfrost_openapi::models::account_withdrawal_content_inner::AccountWithdrawalContentInner;
-use common::pagination::{Pagination, PaginationQuery};
+use common::{
+    accounts::{AccountData, AccountsPath},
+    pagination::{Pagination, PaginationQuery},
+};
 
 pub async fn route(
     Path(path): Path<AccountsPath>,
