+++ src/api/addresses/address/transactions.rs
@@ -1,15 +1,13 @@
-use crate::{
-    BlockfrostError,
-    addresses::{AddressInfo, AddressesPath},
-    api::ApiResult,
-    config::Config,
-};
+use crate::{BlockfrostError, api::ApiResult, config::Config};
 use axum::{
     Extension,
     extract::{Path, Query},
 };
 use blockfrost_openapi::models::address_transactions_content_inner::AddressTransactionsContentInner;
-use common::pagination::{Pagination, PaginationQuery};
+use common::{
+    addresses::{AddressInfo, AddressesPath},
+    pagination::{Pagination, PaginationQuery},
+};
 
 pub async fn route(
     Path(address_path): Path<AddressesPath>,
