+++ src/api/addresses/address/utxos/root.rs
@@ -1,15 +1,13 @@
-use crate::{
-    BlockfrostError,
-    addresses::{AddressInfo, AddressesPath},
-    api::ApiResult,
-    config::Config,
-};
+use crate::{BlockfrostError, api::ApiResult, config::Config};
 use axum::{
     Extension,
     extract::{Path, Query},
 };
 use blockfrost_openapi::models::address_utxo_content_inner::AddressUtxoContentInner;
-use common::pagination::{Pagination, PaginationQuery};
+use common::{
+    addresses::{AddressInfo, AddressesPath},
+    pagination::{Pagination, PaginationQuery},
+};
 
 pub async fn route(
     Path(address_path): Path<AddressesPath>,
