+++ src/api/assets/asset/transactions.rs
@@ -1,11 +1,10 @@
-use crate::{
-    BlockfrostError,
-    api::ApiResult,
-    assets::{AssetData, AssetsPath},
-};
+use crate::{BlockfrostError, api::ApiResult};
 use axum::extract::{Path, Query};
 use blockfrost_openapi::models::asset_transactions_inner::AssetTransactionsInner;
-use common::pagination::{Pagination, PaginationQuery};
+use common::{
+    assets::{AssetData, AssetsPath},
+    pagination::{Pagination, PaginationQuery},
+};
 
 pub async fn route(
     Path(path): Path<AssetsPath>,
