+++ src/lib.rs
@@ -1,19 +1,13 @@
-pub mod accounts;
-pub mod addresses;
 pub mod api;
-pub mod assets;
 pub mod cbor;
 pub mod cli;
 pub mod config;
-pub mod genesis;
 pub mod health_monitor;
 pub mod icebreakers_api;
 pub mod load_balancer;
 pub mod middlewares;
 pub mod node;
-pub mod payment_cred;
 pub mod server;
-pub mod types;
 
 pub use common::errors::{AppError, BlockfrostError};
 pub use node::pool::NodePool;
