+++ src/server/state.rs
@@ -1,7 +1,7 @@
 use crate::config::Config;
-use crate::config::Network;
 use axum::extract::State;
 use blockfrost_openapi::models::genesis_content::GenesisContent;
+use common::types::Network;
 use std::sync::Arc;
 
 #[derive(Clone)]
