--- src/types.rs
@@ -1,7 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
-pub struct Amount {
-    pub unit: String,
-    pub quantity: String,
-}
